chore: clean up hooks, rm active sub tracking
