+++ src/components/Nav/Menu/useSubMenu.ts
@@ -1,4 +1,3 @@
-import { useState } from "react"
 import type { MotionProps } from "framer-motion"
 import { useRouter } from "next/router"
 
@@ -11,25 +10,17 @@ export const useSubMenu = () => {
   const { asPath, locale } = useRouter()
   const menuColors = useNavMenuColors()
   const { isRtl } = useRtlFlip()
-  const [activeSub, setActiveSub] = useState<string | null>(null)
 
   const menuVariants: MotionProps["variants"] = {
     closed: { opacity: 0, scaleX: 0.8, originX: isRtl ? 1 : 0 },
     open: { opacity: 1, scaleX: 1, originX: isRtl ? 1 : 0 },
   }
 
-  const handleSubMenuChange = (value: string) => {
-    setActiveSub(value)
-  }
-
   return {
-    activeSub,
     asPath,
     locale,
     menuColors,
     menuVariants,
     PADDING,
-    setActiveSub,
-    handleSubMenuChange,
   }
 }
